refactor: MarketCard, ProductCard 통합
- ref: #25
