+++ src/mocks/constants/mockMarkets.ts
@@ -1,0 +1,28 @@
+const MOCK_MARKETS = [
+  {
+    id: 0,
+    marketName: '세찬 베이커리',
+    rate: 4.9,
+    reviewer: 196,
+    thumbnailUrl:
+      'https://images.unsplash.com/photo-1581683705068-ca8f49fc7f45?q=80&w=3000&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+  },
+  {
+    id: 1,
+    marketName: '상윤이네 쿠키',
+    rate: 4.8,
+    reviewer: 214,
+    thumbnailUrl:
+      'https://images.unsplash.com/photo-1696189271615-4d484e3fe674?q=80&w=3024&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+  },
+  {
+    id: 2,
+    marketName: '강지원 명인 한과',
+    rate: 5.0,
+    reviewer: 59,
+    thumbnailUrl:
+      'https://images.unsplash.com/photo-1647168587041-ee346e60ad43?q=80&w=3087&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+  },
+];
+
+export default MOCK_MARKETS;
