+++ src/mocks/constants/mockProducts.ts
@@ -1,0 +1,55 @@
+const MOCK_PRODUCTS = [
+  {
+    id: 0,
+    marketName: '세찬 베이커리',
+    productName: '알록달록 마카롱',
+    price: 18000,
+    rate: 4.8,
+    reviewer: 196,
+    thumbnailUrl:
+      'https://images.unsplash.com/photo-1558326567-98ae2405596b?q=80&w=2718&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+  },
+  {
+    id: 1,
+    marketName: '상윤이네 쿠키',
+    productName: '바삭바삭 쿠키 (6개입)',
+    price: 20000,
+    rate: 4.9,
+    reviewer: 214,
+    thumbnailUrl:
+      'https://images.unsplash.com/photo-1618384836169-8471703f1eaf?q=80&w=2680&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D',
+  },
+  {
+    id: 2,
+    marketName: '강지원 명인 한과',
+    productName: '쫀득 수제 약과',
+    price: 11900,
+    rate: 4.8,
+    reviewer: 59,
+    thumbnailUrl:
+      'https://img.freepik.com/premium-photo/close-up-food-plate-table_1048944-17640647.jpg?w=900',
+  },
+  {
+    id: 3,
+    marketName: '지젤 (지훈이가 만든 젤리)',
+    productName: '폭신폭신 젤리',
+    price: 8900,
+    discountPrice: 7900,
+    rate: 4.7,
+    reviewer: 43,
+    thumbnailUrl:
+      'https://img.freepik.com/free-photo/delicious-candy-with-orange-background_23-2149722452.jpg?t=st=1738280652~exp=1738284252~hmac=b5de424639239a506bf2a7cda6d33560e098d6a8f3f0a8bb8cc723680c3d49e4&w=900',
+  },
+  {
+    id: 4,
+    thumbnailUrl: '',
+    marketName: '테스트 마켓',
+    productName: '테스트 상품',
+    price: 10000,
+    discountPrice: 8000,
+    rate: 4.5,
+    reviewer: 100,
+  },
+];
+
+export default MOCK_PRODUCTS;
