+++ src/types/PreviewCard.ts
@@ -1,0 +1,16 @@
+export type CardType = 'market' | 'product';
+
+export type CardSize = 'small' | 'medium' | 'large';
+
+export interface PreviewCardProps {
+  id: number;
+  marketName?: string;
+  rate: number;
+  reviewer: number;
+  productName?: string;
+  price?: number;
+  discountPrice?: number;
+  thumbnailUrl: string;
+  type: CardType;
+  size?: CardSize;
+}
